+++ Gruntfile.js
@@ -21,6 +21,9 @@ module.exports = function (grunt) {
         tangelo_dir = path.resolve(lib + (windows ? "" : "python-2.7/" + "site-packages/tangelo")),
         version = grunt.file.readJSON("package.json").version,
         tangeloCmdLine,
+        pkgDataDir = "tangelo/tangelo/pkgdata/",
+        pluginDir = pkgDataDir + "plugin/",
+        webDir = pkgDataDir + "web/",
         styleCheckFiles;
 
     tangeloCmdLine = function (hostname, port, root, cover) {
@@ -40,7 +43,7 @@ module.exports = function (grunt) {
                 "vtkweb",
                 "vtkweb/web"
             ].map(function (p) {
-                return "venv/share/tangelo/plugin/" + p;
+                return "venv/lib/python2.7/site-packages/tangelo/pkgdata/plugin/" + p;
             });
         }
 
@@ -58,21 +61,21 @@ module.exports = function (grunt) {
                 "--host", hostname,
                 "--port", port,
                 "--root", root,
-                "--plugin-config", "venv/share/tangelo/plugin/plugin.conf"
+                "--config", "tests/bundled-plugins.yaml"
             ])
         };
     };
 
     styleCheckFiles = [
         "js/src/**/*.js",
-        "tangelo/plugin/**/*.js",
-        "!tangelo/plugin/docs/**/*.js",
-        "!tangelo/plugin/**/geo.min.js",
-        "!tangelo/plugin/**/geo.ext.min.js",
-        "!tangelo/plugin/**/vgl.min.js",
-        "!tangelo/plugin/tangelo/web/tangelo.min.js",
-        "!tangelo/plugin/vtkweb/web/lib/autobahn.min.js",
-        "!tangelo/plugin/vtkweb/web/lib/vtkweb-all.min.js"
+        pluginDir + "**/*.js",
+        "!" + pluginDir + "docs/**/*.js",
+        "!" + pluginDir + "**/geo.min.js",
+        "!" + pluginDir + "**/geo.ext.min.js",
+        "!" + pluginDir + "**/vgl.min.js",
+        "!" + pluginDir + "tangelo/web/tangelo.min.js",
+        "!" + pluginDir + "vtkweb/web/lib/autobahn.min.js",
+        "!" + pluginDir + "vtkweb/web/lib/vtkweb-all.min.js"
     ];
 
     // Project configuration.
@@ -80,7 +83,7 @@ module.exports = function (grunt) {
       version: {
           src: [
               "tangelo/tangelo/__main__.py",
-              "tangelo/plugin/tangelo/web/version.py",
+              pluginDir + "tangelo/web/version.py",
               "tangelo/setup.py",
               "js/src/core.js"
           ]
@@ -92,7 +95,7 @@ module.exports = function (grunt) {
           },
           dist: {
               src: ["js/src/**/*.js"],
-              dest: "tangelo/plugin/tangelo/web/tangelo.js"
+              dest: pluginDir + "tangelo/web/tangelo.js"
           }
       },
       uglify: {
@@ -101,7 +104,7 @@ module.exports = function (grunt) {
           },
           dist: {
               src: "<%= concat.dist.dest %>",
-              dest: "tangelo/plugin/tangelo/web/tangelo.min.js"
+              dest: pluginDir + "tangelo/web/tangelo.min.js"
           }
       },
       jshint: {
@@ -199,8 +202,8 @@ module.exports = function (grunt) {
           package: [
               "tangelo/MANIFEST",
               "tangelo/README",
-              "tangelo/plugin/docs",
-              "tangelo/web/js"
+              pluginDir + "docs",
+              webDir + "js"
           ]
       }
     });
@@ -379,13 +382,21 @@ module.exports = function (grunt) {
 
     // Install the Python package to the virtual environment.
     grunt.registerTask("install", "Install Tangelo to the virtual environment", function () {
-        var done;
+        var done,
+            pyversion;
 
         done = this.async();
 
+        // This is necessary to reconcile Python setuptools's notion of version
+        // numbers with npm's.  Both accept "foobar-0.8.1-dev" as a valid
+        // version number, but setuptools will "normalize" it to
+        // "foobar-0.8.1.dev0", so we need to do the same in order to install
+        // the package created by the grunt package task.
+        pyversion = version.replace("-dev", ".dev0");
+
         grunt.util.spawn({
             cmd: pip,
-            args: ["install", "--upgrade", "sdist/tangelo-" + version + zipExt],
+            args: ["install", "--upgrade", "sdist/tangelo-" + pyversion + zipExt],
             opts: {
                 stdio: "inherit"
             }
@@ -517,7 +528,7 @@ module.exports = function (grunt) {
                    "-D", "version=" + version,
                    "-D", "release=" + version,
                    "docs",
-                   "tangelo/plugin/docs/web"],
+                   pluginDir + "docs/web"],
             opts: {
                 stdio: "inherit"
             }
@@ -543,7 +554,7 @@ module.exports = function (grunt) {
             host = "localhost";
         }
 
-        tangeloCmd = tangeloCmdLine(host, port, "venv/share/tangelo/web", false);
+        tangeloCmd = tangeloCmdLine(host, port, "venv/lib/python2.7/site-packages/tangelo/pkgdata/web", false);
 
         grunt.util.spawn({
             cmd: tangeloCmd.cmd,